pyproject.toml
[tool.poetry]
name = "twitch-desktop"
version = "0.1.0"
description = ""
authors = ["Don Nguyen <ndon55555@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGObject = "^3.38.0"
python-vlc = "^3.0.11115"
twitch-python = "^0.0.19"
oauthlib = "^3.1.0"
Flask = "^1.1.2"
selenium = "^3.141.0"
geckodriver-autoinstaller = "^0.1.0"
Authlib = "^0.15.2"
requests = "^2.25.1"
twitchio = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry_core>=1.0"]
build-backend = "poetry.core.masonry.api"
